backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
cryptography==43.0.1
python-multipart==0.0.12
redis==5.2.0
psycopg2-binary==2.9.10
alembic==1.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
httpx==0.27.2
aiohttp==3.11.4
bcrypt==4.1.2
pgvector==0.2.5
prometheus-client==0.19.0
websockets==12.0

